refactor: Use `@guardian/cdk` mock to ease version upgrades
It is currently impossible to use Dependabot to update `@guardian/cdk`.

Dependabot will raise a PR to update the version of `@guardian/cdk` in `package.json` and nothing else.

This PR will always fail CI as the value of the `gu:cdk:version` tag in our snapshots is tied to the version of `@guardian/cdk` being used.

That is, if Dependabot raises a PR to move from v1 to v2, as the snaphots are not updated at the same time, the tests fail.

In this change, we use the newly published mock in `@guardian/cdk` which results in the value of `gu:cdk:version` in the snapshots being static and thus Dependabot PRs become usable and we update `@guardian/cdk` more frequently.

This does not affect the production template generated by `cdk synth`.

See:
 - https://github.com/guardian/cdk/pull/448