feat: Collect and expose lambda package type
The package type is:

> The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

If the package type is `"Image"` then the lambda runtime is `null`.
Exposing package type provides some context to a `null` runtime.

This requires an update of the AWS SDK (2.16.58 is the current latest).
To support this, we have to add `jackson-module-scala` to solve the
issue of:

```
com.fasterxml.jackson.databind.JsonMappingException: Scala module 2.10.5 requires Jackson Databind version >= 2.10.0 and < 2.11.0
```

I don't fully understand how this solves things...

Pasting random code from the internet has served me well so far 😬.

See:
  - https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-function.html#cfn-lambda-function-packagetype
  - https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-function.html#cfn-lambda-function-runtime
  - https://github.com/FasterXML/jackson-module-scala/issues/513#issuecomment-812328540