chore: Run Dependabot monthly
The CDK project in this repository doesn't run in a PROD environment, only CI, so we can afford to use old versions of libraries for a short time. Update Dependabot's configuration to run monthly to reduce the frequency of PRs by a quarter and thus require less commitment from the team.