fix: Keep account number if already provided in config
Prism's collectors operate in various AWS accounts.

AWS accounts are listed in two ways within configuration:
  1. The name, and the IAM role to assume
  2. The name, the IAM role to assume, and the account number

Accounts of the second form are not Guardian accounts, but AWS ones.
The IAM role provided for these is the role of the deployTools account.

That is, we use deployTools credentials to query these accounts.
The data in the AWS accounts are publicly accessible, however the SDK still needs credentials.

In this change, we correct the logic that merges these two lists into one.

Previously the logic was:
  1. Merge the two lists
  2. For each item, make a `getCallerIdentity` request to find the account ID
  3. Update the item, setting the account ID

However, for the second category of account, this is incorrect as the value from step
2 will incorrectly evaluate to the deployTools account.

In this change, if the account number has already been provided, we do not perform the `getCallerIdentity` request,
retaining the value instead.