fix: Resolve typescript dependency issue
Since updating @guardian/eslint-config-typescript to v3.0.0,
we're seeing this issue when installing dependencies with yarn:

```
error An unexpected error occurred: "could not find a copy of typescript to link in /cdk/node_modules/@guardian/eslint-config-typescript/node_modules".
```

In this change we fix this by switching to npm.

Given we're only using `yarn` for CDK in this project, swapping to `npm` is fairly innocent.